fix jwt

+++ resources/item.py
@@ -1,6 +1,5 @@
-
-from flask_restful import Resource, reqparse, request
-from flask_jwt import jwt_required
+from flask_restful import Resource, reqparse
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
 from models.item import ItemModel
 
 class Item(Resource):
@@ -16,7 +15,7 @@ class Item(Resource):
         if item:
             return item.json()
         return  {'message': 'item not found'}, 404
-           
+    @jwt_required(fresh=True)  
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': 'in item with {} name already exists'.format(name)}, 400
@@ -30,12 +29,17 @@ class Item(Resource):
         except:
             return  {'message': 'an error occurred'}, 500
         return item.json(), 201
-
+    @jwt_required()
     def delete(self, name):
+        claims=get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privelege required'}, 401
+
         item=ItemModel.find_by_name(name)
         if item is not None:
             item.delete_from_db()
-        return {"message": "item deleted"}
+            return {"message": "item deleted"}
+        return {"message": "item not found"}
         
     def put(self,name):
         data = Item.parser.parse_args()
@@ -57,6 +61,12 @@ class Item(Resource):
     
 
 class ItemList(Resource):
+    @jwt_required(optional=True)
     def get(self):
+        user_id=get_jwt_identity()
+        items=[item.json() for item in ItemModel.find_all()]
         
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [item['name'] for item in items], 'message': 'more info available if you log in'}, 200
+
